split media related migrations into service dir
